pyproject.toml
[project]
name = "upload-tunkin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pymysql-pool==0.5.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqids>=0.5.2",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.2.250926",
]
